adding user management JS file

+++ allLogin.js
@@ -5,7 +5,6 @@ loginPage.addEventListener('click', (e) => {
     window.location.href = './userReg.html';
 });
 
-
 function validate(e){
     e.preventDefault();
     let adminName = document.getElementById('adminname').value;
@@ -24,17 +23,25 @@ function register(e) {
     e.preventDefault();
     
     let formData = JSON.parse(localStorage.getItem('formData')) || [];
-    formData.push({
-        userName : document.getElementById('userName').value,
-        userEmail : document.getElementById('userEmail').value,
-        userPassword : document.getElementById('userPass').value,
-        userMobileNumber : document.getElementById('mobile').value,
-    });
-    localStorage.setItem('formData', JSON.stringify(formData));
-    document.querySelector('form').reset();
+    let exist = formData.length && JSON.parse(localStorage.getItem('formData')).some(data => data.userName.toLowerCase() == document.getElementById('userName').value.toLowerCase() && data.userEmail.toLowerCase() == document.getElementById('userEmail').value.toLowerCase());
+
+    if(!exist){
+        formData.push({
+            userName : document.getElementById('userName').value,
+            userEmail : document.getElementById('userEmail').value,
+            userPassword : document.getElementById('userPass').value,
+            userMobileNumber : document.getElementById('mobile').value,
+            id : new Date().getTime(),
+            status : 'Inactive',                    
+        });
+        localStorage.setItem('formData', JSON.stringify(formData));
+        document.querySelector('form').reset();
+        alert('User Registered Successfully');
+    }
+    else{
+        alert('User Already Exist');
+    }
 }
-
-
 function check(e){
     e.preventDefault();
     let userName = document.getElementById('uName').value;
@@ -42,47 +49,69 @@ function check(e){
     let formData = JSON.parse(localStorage.getItem('formData'));
     for(let i = 0; i < formData.length; i++){
         if((formData[i].userName == userName) && (formData[i].userPassword == userPassword)){
-            window.location.href = './userPage.html';
-            return true;
+            if(formData[i].status == "Active"){
+                localStorage.setItem('userName', JSON.stringify(formData[i].userName));
+                localStorage.setItem('userEmail', JSON.stringify(formData[i].userEmail));
+                localStorage.setItem('userMobileNumber', JSON.stringify(formData[i].userMobileNumber));
+                window.location.href = './userPage.html';
+            }
+            else if(formData[i].status == "Inactive"){
+                alert('Your Account is Inactive');
+            }
         }
     }
 }
 
+function handleChange(id){
+    let button = document.getElementById(id);
+    let data = JSON.parse(localStorage.getItem('formData'));
+        button.addEventListener('click', function(){
+            if(button.id == id){
+                for(let i = 0; i < data.length; i++){
+                    if(data[i].id == id){
+                       if(data[i].status == 'Inactive'){
+                           data[i].status = 'Active';
+                            button.innerHTML = 'Active';
+                       }
+                       else{
+                           data[i].status = 'Inactive';
+                           button.innerHTML = 'Inactive';
+                       }
+                    }
+                }
+                localStorage.setItem('formData', JSON.stringify(data));
+            }
+            
+        });
+}
+
 function userData(){
-    console.log(localStorage.getItem('formData'));
-   if(localStorage.getItem('formData')){
-       var output = document.querySelector('tbody');
-       output.innerHTML = "";
-       JSON.parse(localStorage.getItem('formData')).forEach(data => {
-           output.innerHTML += `
+    let formData = JSON.parse(localStorage.getItem('formData'));
+    let output = document.querySelector('tbody');
+    output.innerHTML = "";
+    formData.forEach(data => {
+        output.innerHTML += `
            <tr>
               <td>${data.userName}</td>
               <td>${data.userEmail}</td>
               <td>${data.userPassword}</td>
               <td>${data.userMobileNumber}</td>
-              <td><label class="switch">
-              <input type="checkbox">
-              <span class="slider"></span>
-            </label></td>
-            </tr>`;
-       })
-   }
+              <td>
+              <button id="${data.id}" class="bton" onclick="handleChange(${data.id})">${data.status}</button>
+              </td>
+            </tr>
+        `;
+    })
 }
 
 function displayUserDetails(){
-    var userName = document.getElementById('uName').value;
-    var userPassword = document.getElementById('uPass').value;
-    // console.log(userName);
-    let div = document.querySelector('tbody');
-    // div.innerHTML = "";
-    let formData = JSON.parse(localStorage.getItem('formData'));
-    // console.log(formData);
-   
-    formData.forEach(data => {
-        if((data.userName == userName) && (data.userPassword == userPassword)){
-           console.log(data);
-           console.log(userName);
-           console.log(data.userPassword);
-        }
-    })
+   let userName = JSON.parse(localStorage.getItem('userName'));
+   let userEmail = JSON.parse(localStorage.getItem('userEmail'));
+   let userMobileNumber = JSON.parse(localStorage.getItem('userMobileNumber'));
+   let details = document.getElementById('details');
+   details.innerHTML = `
+   <h2>Hello ${userName.toUpperCase()}!!</h2>
+   <h3>Your Email is  <span>${userEmail}<span></h3>
+   <h3>Your Mobile Number is  <span>${userMobileNumber}</span></h3>`;
 }
+//
